requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiolimiter==1.2.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
baidusearch==1.0.3
beartype==0.12.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.37.38
botocore==1.37.38
browser-use==0.1.41
browsergym==0.13.3
browsergym-assistantbench==0.13.3
browsergym-core==0.13.3
browsergym-experiments==0.13.3
browsergym-miniwob==0.13.3
browsergym-visualwebarena==0.13.3
browsergym-webarena==0.13.3
browsergym-workarena==0.4.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
cython==3.0.12
dataclasses-json==0.6.7
datasets==3.5.0
debugpy==1.8.14
decorator==5.2.1
deepl==1.21.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
duckduckgo-search==8.0.1
english-words==2.0.1
evaluate==0.4.3
executing==2.2.0
faiss-cpu==1.10.0
faker==37.1.0
farama-notifications==0.0.4
fastapi==0.115.12
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
firecrawl-py==2.5.3
flask==3.1.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
google-genai==1.11.0
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
gotrue==2.12.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
gymnasium==1.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.14
kiwipiepy==0.20.4
kiwipiepy-model==0.20.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-community==0.3.20
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-mcp-adapters==0.0.9
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-teddynote==0.3.45
langchain-text-splitters==0.3.7
langgraph==0.3.34
langgraph-checkpoint==2.0.25
langgraph-checkpoint-sqlite==2.0.7
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.33
lazy-loader==0.4
libvisualwebarena==0.0.15
libwebarena==0.0.4
loguru==0.7.3
lxml==5.3.2
lz4==4.4.4
markdown-it-py==3.0.0
markdownify==1.1.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mem0ai==0.1.88
mistune==3.1.3
mmh3==3.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.0
notebook-shim==0.2.4
numpy==2.2.5
olefile==0.47
ollama==0.4.8
openai==1.75.0
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.2.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.5.4
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
portalocker==2.10.1
postgrest==1.0.1
posthog==3.25.0
primp==0.15.0
prometheus-client==0.21.1
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
protoc-gen-openapiv2==0.0.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.9.1
pyee==12.1.1
pygments==2.19.1
pyjwt==2.10.1
pyobjc-core==11.0
pyobjc-framework-cocoa==11.0
pyparsing==3.2.3
pyperclip==1.9.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-httpx==0.35.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.4.0
qdrant-client==1.13.3
rank-bm25==0.2.2
realtime==2.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.11.5
sacrebleu==2.5.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
send2trash==1.8.3
sentence-transformers==4.1.0
setuptools==79.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.2
storage3==0.11.3
strenum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.1
tabulate==0.9.0
tavily-python==0.7.0
tenacity==9.1.2
terminado==0.18.1
text-generation==0.7.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.3
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250328
types-tqdm==4.67.0.20250417
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
unidiff==0.7.5
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
weblinx==0.3.2
weblinx-browsergym==0.0.2
websocket-client==1.8.0
websockets==14.2
werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
